test(wasap): merge Collection analysis mode test into wastewater.spec.ts
